+++ anne_auth/app/controllers/anne_auth/accounts/email_verifications_controller.rb
@@ -17,10 +17,12 @@ module AnneAuth
         code_lookup = verification_token_class.lookup_for(current_account, params[:otp])
 
         if code_lookup.success?
-          verify_account(code_lookup.verification_token)
-          redirect_to auth_route(:root_path), notice: "メール認証が完了しました。"
+          account = verify_account(code_lookup.verification_token)
+          redirect_to after_account_email_verification_url(account), status: :see_other, notice: "メール認証が完了しました。"
         else
-          redirect_to auth_route(:account_email_verification_pending_path), alert: verification_error_message(code_lookup.status)
+          redirect_to auth_route(:account_email_verification_pending_path),
+            status: :see_other,
+            alert: verification_error_message(code_lookup.status)
         end
       end
 
@@ -32,18 +34,27 @@ module AnneAuth
         _verification_token, plain_code = verification_token_class.issue_for(current_account)
         AnneAuth.configuration.account_mailer_class.with(account: current_account, plain_code:).verification.deliver_now
 
-        redirect_to auth_route(:account_email_verification_pending_path), notice: "認証メールを再送しました。"
+        redirect_to auth_route(:account_email_verification_pending_path),
+          status: :see_other,
+          notice: "認証メールを再送しました。"
       end
 
       private
+        def after_account_email_verification_url(account)
+          AnneAuth.configuration.after_account_email_verification_path.call(self, account)
+        end
+
         def verification_token_class
           AnneAuth.configuration.account_verification_token_class
         end
 
         def verify_account(verification_token)
           verification_token.verify!
+          account = verification_token.account
 
-          start_new_account_session_for(verification_token.account) unless account_authenticated?
+          start_new_account_session_for(account) unless account_authenticated?
+
+          account
         end
 
         def verification_error_message(status)